+++ src/test/http/walk.http
@@ -1,7 +1,7 @@
 # Walk 테스트
 
 ### AccessToken
-@accessToken = eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJwZXRtaWx5IiwiaWF0IjoxNzI0MDYxNDczLCJleHAiOjE3MjQwNjUwNzMsInN1YiI6ImFiYzEifQ.pCZhaWFJZrGIj2biCJl0adLxFtspplca7AKIIvyygCs
+@accessToken = eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJwZXRtaWx5IiwiaWF0IjoxNzI0MTI3MDYyLCJleHAiOjE3MjQxMzA2NjIsInN1YiI6ImFiYzEifQ.alZ-LZ-viFXmfEDa-jaqjXhKGvZIBsNeB4agaTobIbg
 
 ### 산책 시작
 POST http://localhost:8080/api/v1/walks
