+++ src/test/java/com/concord/petmily/walk/WalkControllerTest.java
@@ -1,7 +1,7 @@
 package com.concord.petmily.walk;
 
 import com.concord.petmily.walk.controller.WalkController;
-import com.concord.petmily.walk.service.WalkService;
+import com.concord.petmily.walk.service.WalkServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -18,7 +18,7 @@ public class WalkControllerTest {
     private MockMvc mockMvc;
 
     @MockBean
-    private WalkService walkService;
+    private WalkServiceImpl walkServiceImpl;
 
     @Test
     @WithMockUser(username = "user1")
